Convert hiveconf and hivetool to Python 3
Hiveconf now always operates in UTF-8. This means that if you ship .hconf-files
they must be encoded with UTF-8.

If files contain any non-ASCII characters that has other encoding than
UTF-8, Python will throw a UnicodeDecodeError. In order to give a more
informative error message, Hiveconf now has an additional exception
class to handle this type of error. Note however that we do not catch
the errors. This is in line with how Hiveconf handles syntax errors in
files.

Signed-off-by: Frida Flodin <frifl@cendio.se>