Move function walk() from hiveconf to hivetool
This function is only used by hivetool, so there's no need to clutter up
hiveconf with it. Hiveconf class _IndentPrinter was only used by walk(),
so this class was also moved to hivetool.

Also renamed the walk() function to print_walk(). There could be several
walk functions, so give this a more specific name so we can tell them
more easily apart.

Signed-off-by: Frida Flodin <frifl@cendio.se>